third_party/blink/web_tests/svg/parser/whitespace-angle-invalid-1-expected.txt

CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaN".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnord".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnorddeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaN ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnord ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnorddeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaN   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnord   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnorddeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaN\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitygrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitygrad\r\n\t…".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnord\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnorddeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaN\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "NaNgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnord\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnorddeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "fnordgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "Egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "E-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaN".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnord".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnorddeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaN ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnord ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnorddeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaN   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnord   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnorddeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaN\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitygrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitygrad\r\n\t…".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnord\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnorddeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaN\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " NaNgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnord\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnorddeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " fnordgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " Egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " E-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaN".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnord".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnorddeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaN ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnord ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnorddeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaN   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnord   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnorddeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaN\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitygrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitygrad\r\n\t…".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnord\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnorddeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaN\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   NaNgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnord\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnorddeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   fnordgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   Egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   E-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaN".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnord".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnorddeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t edeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t erad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t egrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaN ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnord ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnorddeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t edeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t erad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t egrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaN   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnord   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnorddeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t edeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t erad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t egrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaN\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitygrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitygrad\r\n\t…".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnord\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnorddeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t edeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t erad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t egrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaN\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t NaNgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnord\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnorddeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t fnordgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t Egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t edeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t erad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t egrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t E-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaN".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinity".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitydeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinityrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitygrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnord".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnorddeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEdeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fErad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEgrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fedeg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ferad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fegrad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaN ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinity ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitydeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinityrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitygrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnord ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnorddeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEdeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fErad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEgrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fedeg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ferad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fegrad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaN   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinity   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitydeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinityrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitygrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnord   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnorddeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEdeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fErad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEgrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fedeg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ferad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fegrad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaN\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitygrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinity\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitydeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinityrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitygrad\r\n\t…".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnord\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnorddeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEdeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fErad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEgrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fedeg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ferad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fegrad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaN\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fNaNgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fInfinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinity\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitydeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinityrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-Infinitygrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnord\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnorddeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ffnordgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEdeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fErad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fEgrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fedeg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\ferad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fegrad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fE-grad\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
